@@ -12,46 +12,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.2",
-    "@angular/cdk": "7.2.2",
-    "@angular/common": "7.2.2",
-    "@angular/compiler": "7.2.2",
-    "@angular/core": "7.2.2",
-    "@angular/forms": "7.2.2",
-    "@angular/http": "7.2.2",
-    "@angular/material": "7.2.2",
-    "@angular/platform-browser": "7.2.2",
-    "@angular/platform-browser-dynamic": "7.2.2",
-    "@angular/router": "7.2.2",
-    "@angular/upgrade": "7.2.2",
+    "@angular/animations": "8.0.0",
+    "@angular/cdk": "8.0.0",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
+    "@angular/material": "8.0.0",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/router": "8.0.0",
+    "@angular/upgrade": "8.0.0",
     "classlist.js": "^1.1.20150312",
-    "core-js": "2.5.7",
+    "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
-    "jasmine": "^3.1.0",
-    "rxjs": "6.3.3",
+    "jasmine": "^3.4.0",
+    "rxjs": "6.5.2",
     "tslib": "^1.9.3",
     "web-animations-js": "^2.3.1",
-    "zone.js": "0.8.29"
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.3",
-    "@angular/cli": "7.2.2",
-    "@angular/compiler-cli": "7.2.2",
-    "@angular/language-service": "7.2.2",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "~0.800.2",
+    "@angular/cli": "8.0.2",
+    "@angular/compiler-cli": "8.0.0",
+    "@angular/language-service": "8.0.0",
+    "@types/jasmine": "^3.3.13",
+    "@types/jasminewd2": "^2.0.3",
     "@types/node": "^10.12.18",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "codelyzer": "~5.1.0",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.0",
-    "ts-node": "^7.0.0",
-    "tslint": "^5.11.0",
-    "typescript": "~3.2.2"
+    "protractor": "^5.4.2",
+    "ts-node": "^8.2.0",
+    "tslint": "^5.17.0",
+    "typescript": "~3.4.2"
   }
 }
