@@ -1,7 +1,10 @@
 import { Component } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { Employee, EmployeeLoaderService } from './employee-loader.service';
+import {
+  Employee,
+  EmployeeLoaderService
+} from './employee-loader.service';
 
 @Component({
   selector: 'app-root',
